fix: auth user can not visit login & signup page
